get current user from userManager

+++ src/main/java/model/team/TeamManager.java
@@ -1,9 +1,11 @@
 package model.team;
 
 import model.UnauthorisedOperationException;
+import model.user.NoSignedInUserException;
 import model.user.User;
 import model.team.repository.TeamRepository;
 import model.team.repository.TeamRepositoryFactory;
+import model.user.UserManager;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -34,11 +36,14 @@ public class TeamManager {
    *
    * @param name is the name of the new team.
    * @throws SQLException if the operation could not be performed in the database.
+   * @throws NoSignedInUserException if the user is not signed in.
    */
-  public void createNewTeam(String name) throws SQLException {
-    // todo: get managerId from UserManager
-    int managerId = 2;
-    teamRepository.saveTeam(new Team(name, managerId, generateTeamCode()));
+  public void createNewTeam(String name) throws SQLException, NoSignedInUserException {
+    if (UserManager.getInstance().getCurrentUser().isEmpty()) {
+      throw new NoSignedInUserException("create new team");
+    }
+    User currentUser = UserManager.getInstance().getCurrentUser().get();
+    teamRepository.saveTeam(new Team(name, currentUser.getId().get(), generateTeamCode()));
   }
 
   /**
@@ -50,14 +55,18 @@ public class TeamManager {
    * @throws InexistentTeamException if no team with teamId exists in the database.
    * @throws UnauthorisedOperationException if the current user is not allowed to delete the team,
    *     because the user is not the manager of the team.
+   * @throws NoSignedInUserException if the user is not signed in.
    */
   public void deleteTeam(int teamId)
-      throws SQLException, InexistentTeamException, UnauthorisedOperationException {
+          throws SQLException, InexistentTeamException, UnauthorisedOperationException, NoSignedInUserException {
     Optional<Team> team = teamRepository.getTeam(teamId);
     if (team.isEmpty()) {
       throw new InexistentTeamException(teamId);
     }
-    User currentUser = new User(1, "", ""); // todo get from UserManager
+    if (UserManager.getInstance().getCurrentUser().isEmpty()) {
+      throw new NoSignedInUserException("delete team");
+    }
+    User currentUser = UserManager.getInstance().getCurrentUser().get();
     if (currentUser.getId().isEmpty() || team.get().getManagerId() != currentUser.getId().get()) {
       throw new UnauthorisedOperationException(
           currentUser.getId().get(),
@@ -72,20 +81,25 @@ public class TeamManager {
    *
    * @return the list of teams having the current user as member.
    * @throws SQLException if the operation could not be performed in the database.
+   * @throws NoSignedInUserException if the user is not signed in.
    */
-  public List<Team> getTeamsOfCurrentUser() throws SQLException {
-    User currentUser = new User(1, "", ""); // todo get from UserManager
+  public List<Team> getTeamsOfCurrentUser() throws SQLException, NoSignedInUserException {
+    if (UserManager.getInstance().getCurrentUser().isEmpty()) {
+      throw new NoSignedInUserException("see teams");
+    }
+    User currentUser = UserManager.getInstance().getCurrentUser().get();
     return teamRepository.getTeamsOfUser(currentUser.getId().get());
   }
 
   /**
    * Generated a new, unique code for the team with the given id, and saves it. Remark that if no
-   * team with the given id exists, nothing happend, no exception is thrown.
+   * team with the given id exists, nothing happens, no exception is thrown.
    *
    * @param teamId is the id of the team for which the new code is generated.
    * @throws SQLException if the operation could not be performed in the database.
    */
   public String regenerateTeamCode(int teamId) throws SQLException {
+    // todo check if the user is the manager
     String newCode = generateTeamCode();
     teamRepository.setNewCode(teamId, newCode);
     return newCode;
@@ -97,10 +111,14 @@ public class TeamManager {
    * @param code is the code provided by the user of the team to join.
    * @throws SQLException if the operation could not be performed in the database.
    * @throws InexistentTeamException if no team with this code exists.
+   * @throws NoSignedInUserException if the user is not signed in.
    */
-  public void joinTeam(String code) throws SQLException, InexistentTeamException {
+  public void joinTeam(String code) throws SQLException, InexistentTeamException, NoSignedInUserException {
     Optional<Team> team = teamRepository.getTeam(code);
-    User currentUser = new User(2, "", ""); // todo get from UserManager
+    if (UserManager.getInstance().getCurrentUser().isEmpty()) {
+      throw new NoSignedInUserException("join team");
+    }
+    User currentUser = UserManager.getInstance().getCurrentUser().get();
     if (team.isEmpty()) {
       throw new InexistentTeamException(code);
     } else {
@@ -114,9 +132,13 @@ public class TeamManager {
    *
    * @param teamId is the id of the team to join.
    * @throws SQLException if the operation could not be performed in the database.
+   * @throws NoSignedInUserException if the user is not signed in.
    */
-  public void leaveTeam(int teamId) throws SQLException {
-    User currentUser = new User(1, "", ""); // todo get from UserManager
+  public void leaveTeam(int teamId) throws SQLException, NoSignedInUserException {
+    if (UserManager.getInstance().getCurrentUser().isEmpty()) {
+      throw new NoSignedInUserException("leave team");
+    }
+    User currentUser = UserManager.getInstance().getCurrentUser().get();
     teamRepository.removeTeamMember(teamId, currentUser.getId().get());
   }
 
@@ -129,18 +151,22 @@ public class TeamManager {
    * @throws SQLException if the operation could not be performed in the database.
    * @throws InexistentTeamException if no team with this teamId exists.
    * @throws UnauthorisedOperationException if the current user is no the manager of the team.
+   * @throws NoSignedInUserException if the user is not signed in.
    */
   public void addMemberToTeam(int teamId, String userName)
-      throws SQLException, InexistentTeamException, UnauthorisedOperationException {
+          throws SQLException, InexistentTeamException, UnauthorisedOperationException, NoSignedInUserException {
     Optional<Team> team = teamRepository.getTeam(teamId);
     if (team.isEmpty()) {
       throw new InexistentTeamException(teamId);
     }
-    User currentUser = new User(1, "", ""); // todo get from UserManager
+    if (UserManager.getInstance().getCurrentUser().isEmpty()) {
+      throw new NoSignedInUserException("add team member");
+    }
+    User currentUser = UserManager.getInstance().getCurrentUser().get();
     if (currentUser.getId().isEmpty() || team.get().getManagerId() != currentUser.getId().get()) {
       throw new UnauthorisedOperationException(
           currentUser.getId().get(),
-          " add a new team to the project",
+          "add a new team to the project",
           "this user is not the manager of the project");
     }
     User newMember = new User(3, userName, ""); // todo get from UserManager
@@ -157,14 +183,18 @@ public class TeamManager {
    * @throws SQLException if the operation could not be performed in the database.
    * @throws InexistentTeamException if no team with this teamId exists.
    * @throws UnauthorisedOperationException if the current user is no the manager of the team.
+   * @throws NoSignedInUserException if the user is not signed in.
    */
   public void removeTeamMember(int teamId, String userName)
-      throws SQLException, InexistentTeamException, UnauthorisedOperationException {
+          throws SQLException, InexistentTeamException, UnauthorisedOperationException, NoSignedInUserException {
     Optional<Team> team = teamRepository.getTeam(teamId);
     if (team.isEmpty()) {
       throw new InexistentTeamException(teamId);
     }
-    User currentUser = new User(1, "", ""); // todo get from UserManager
+    if (UserManager.getInstance().getCurrentUser().isEmpty()) {
+      throw new NoSignedInUserException("remove team member");
+    }
+    User currentUser = UserManager.getInstance().getCurrentUser().get();
     if (currentUser.getId().isEmpty() || team.get().getManagerId() != currentUser.getId().get()) {
       throw new UnauthorisedOperationException(
           currentUser.getId().get(),
@@ -187,14 +217,18 @@ public class TeamManager {
    *     they are not authorised to change the manager.
    * @throws IllegalArgumentException if the user with newManagerName is not the member of the team
    *     in which he/she should become a manager.
+   * @throws NoSignedInUserException if the user is not signed in.
    */
   public void passManagerPosition(int teamId, String newManagerName)
-      throws SQLException, InexistentTeamException, UnauthorisedOperationException {
+          throws SQLException, InexistentTeamException, UnauthorisedOperationException, NoSignedInUserException {
     Optional<Team> team = teamRepository.getTeam(teamId);
     if (team.isEmpty()) {
       throw new InexistentTeamException(teamId);
     }
-    User currentUser = new User(1, "", ""); // todo get from UserManager
+    if (UserManager.getInstance().getCurrentUser().isEmpty()) {
+      throw new NoSignedInUserException("pass manager position of team");
+    }
+    User currentUser = UserManager.getInstance().getCurrentUser().get();
     if (currentUser.getId().isEmpty() || team.get().getManagerId() != currentUser.getId().get()) {
       throw new UnauthorisedOperationException(
           currentUser.getId().get(),
@@ -225,14 +259,18 @@ public class TeamManager {
    * @throws InexistentTeamException if there is no team with id teamId.
    * @throws UnauthorisedOperationException if the current user is not the manager of the team, so
    *     they are not authorised to change the name.
+   * @throws NoSignedInUserException if the user is not signed in.
    */
   public void setNewName(int teamId, String newTeamName)
-      throws SQLException, InexistentTeamException, UnauthorisedOperationException {
+          throws SQLException, InexistentTeamException, UnauthorisedOperationException, NoSignedInUserException {
     Optional<Team> team = teamRepository.getTeam(teamId);
     if (team.isEmpty()) {
       throw new InexistentTeamException(teamId);
     }
-    User currentUser = new User(1, "", ""); // todo get from UserManager
+    if (UserManager.getInstance().getCurrentUser().isEmpty()) {
+      throw new NoSignedInUserException("set new name for team");
+    }
+    User currentUser = UserManager.getInstance().getCurrentUser().get();
     if (currentUser.getId().isEmpty() || team.get().getManagerId() != currentUser.getId().get()) {
       throw new UnauthorisedOperationException(
           currentUser.getId().get(),
